split http filter logic into a fuzzing class
